Refactore switch logic to shared hook
